+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 
+- Inspection: do not use --platform flag with FROM
 - Inspection: [wget without recommended flags](https://protsenko.dev/infrastructure-security/avoid-wget-without-progress-or-quiet/)
 
 ## [2.0.16] 07-02-2026